SecondCommit

+++ src/main/java/com/example/storeapi/model/Produto.java
@@ -8,7 +8,7 @@ import lombok.Data;
 import lombok.NoArgsConstructor;
 
 import javax.persistence.*;
-import java.math.BigDecimal;
+
 
 
 @AllArgsConstructor
@@ -18,13 +18,18 @@ import java.math.BigDecimal;
 @JsonIgnoreProperties({"hibernateLazyInitializer","handler"})
 @NoArgsConstructor
 public class Produto {
+
     @Id
     @GeneratedValue(strategy = GenerationType.AUTO)
     private Long id;
+
     private String none;
-    private BigDecimal valorUnitario;
+
+    private double valorUnitario;
+
     @Builder.Default
     private Boolean disponivel = true;
+
     @ManyToOne
     @JsonIgnore
     private Restaurante restaurante;
